Code refactoring - fixed issue #1

+++ sidebars/admin.php
@@ -10,6 +10,8 @@ if ( $action != '' ) {
 
 	switch ( $action ) {
 		case 'update-sidebar': {
+			check_admin_referer('update-sidebar');
+			
 			$title       = isset( $_REQUEST['sidebar-title'] ) ? $_REQUEST['sidebar-title'] : '';
 			$alias       = isset( $_REQUEST['sidebar-alias'] ) ? sanitize_title( $_REQUEST['sidebar-alias'] ) : '';
 			$description = isset( $_REQUEST['sidebar-description'] ) ? $_REQUEST['sidebar-description'] : '';
@@ -27,6 +29,8 @@ if ( $action != '' ) {
 			break;
 
 		case 'delete-sidebar': {
+			check_admin_referer('delete-sidebar');
+			
 			$alias = isset( $_REQUEST['alias'] ) ? sanitize_title( $_REQUEST['alias'] ) : '';
 			if ( $alias != '' ) {
 				$sidebar_admin->delete_sidebar( $alias );
